requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
requests==2.32.3
pdfplumber==0.11.4
PyMuPDF==1.24.10
python-multipart==0.0.9
langgraph==0.2.14
langchain-google-genai==2.0.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
smtplib-client==0.1.2